requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.4
gspread==5.12.0
oauth2client==4.1.3
nest_asyncio==1.5.8
